Move variance validation to tests
